+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:51132/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:51391/jest?","mongoDBName":"jest"}
