+++ src/infra/db/mongodb/account/account-mongo-repository.ts
@@ -1,4 +1,4 @@
-import { AddAccountRepository } from '../../../../data/protocols/db/add-account-repository'
+import { AddAccountRepository } from '../../../../data/protocols/db/account/add-account-repository'
 import { AddAccountModel } from '../../../../domain/use-cases/add-account'
 import { AccountModel } from '../../../../domain/models/account'
 import { MongoHelper } from '../helpers/mongo-helper'
