+++ src/main/routes/signup-routes.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import { Router } from 'express'
-import { makeSignUpController } from '../factories/signup/signup'
+import { makeSignUpController } from '../factories/signup/signup-factory'
 import { adaptRoute } from '../adapters/express-route-adapter'
 
 export default (router: Router): void => {
