+++ test/account-mongo-repository.spec.ts
@@ -1,5 +1,5 @@
 import { MongoHelper } from '../src/infra/db/mongodb/helpers/mongo-helper'
-import { AccountMongoRepository } from '../src/infra/db/mongodb/account-repository/account'
+import { AccountMongoRepository } from '../src/infra/db/mongodb/account/account-mongo-repository'
 import { Collection } from 'mongodb'
 
 let accountColletion: Collection
