+++ test/login-validation-factory.spec.ts
@@ -1,5 +1,5 @@
-import { makeSignUpValidation } from '../src/main/factories/signup/signup-validation'
-import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
+import { makeLoginValidation } from '../src/main/factories/login/login-validation-factory'
+import { EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
 import { Validation } from '../src/presentation/protocols/validation'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
 
@@ -13,15 +13,14 @@ const makeEmailValidator = (): EmailValidator => {
   }
   return new EmailValidatorStub()
 }
-describe('SignUpValidation Factory', () => {
+describe('LoginValidation Factory', () => {
   test('Should call ValidationComposite with all validations', () => {
-    makeSignUpValidation()
+    makeLoginValidation()
 
     const validations: Validation[] = []
-    for (const field of ['name', 'email', 'password', 'passwordConfirmation']) {
+    for (const field of ['email', 'password']) {
       validations.push(new RequiredFieldValidation(field))
     }
-    validations.push(new CompareFieldsValidation('password', 'passwordConfirmation'))
     validations.push(new EmailValidation('email', makeEmailValidator()))
     expect(ValidationComposite).toHaveBeenCalledWith(validations)
   })
