+++ test/signup-validation-factory.spec.ts
@@ -1,5 +1,5 @@
-import { makeLoginValidation } from '../src/main/factories/login/login-validation'
-import { EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
+import { makeSignUpValidation } from '../src/main/factories/signup/signup-validation-factory'
+import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, ValidationComposite } from '../src/presentation/helpers/validators'
 import { Validation } from '../src/presentation/protocols/validation'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
 
@@ -13,14 +13,15 @@ const makeEmailValidator = (): EmailValidator => {
   }
   return new EmailValidatorStub()
 }
-describe('LoginValidation Factory', () => {
+describe('SignUpValidation Factory', () => {
   test('Should call ValidationComposite with all validations', () => {
-    makeLoginValidation()
+    makeSignUpValidation()
 
     const validations: Validation[] = []
-    for (const field of ['email', 'password']) {
+    for (const field of ['name', 'email', 'password', 'passwordConfirmation']) {
       validations.push(new RequiredFieldValidation(field))
     }
+    validations.push(new CompareFieldsValidation('password', 'passwordConfirmation'))
     validations.push(new EmailValidation('email', makeEmailValidator()))
     expect(ValidationComposite).toHaveBeenCalledWith(validations)
   })
